[x86] support omp on windows, fix precision profiler and relu6 bug (#5506)

+++ lite/kernels/x86/density_prior_box_compute.cc
@@ -60,7 +60,7 @@ void DensityPriorBoxCompute::Run() {
 #ifdef PADDLE_WITH_MKLML
 #pragma omp parallel for reduction(+ : num_priors)
 #endif
-  for (size_t i = 0; i < densities.size(); ++i) {
+  for (int i = 0; i < densities.size(); ++i) {
     num_priors += (fixed_ratios.size()) * (pow(densities[i], 2));
   }
